+++ easybatis-core/src/test/java/com/xwc/open/easybatis/base/insert/MixtureBaseMapper.java
@@ -24,16 +24,16 @@ public interface MixtureBaseMapper<E, K> extends BaseMapper<E, K> {
     E selectKey(K id);
 
     @UpdateSql
-    Long update(E entity);
+    Integer update(E entity);
 
     @InsertSql
-    Long insert(E entity);
+    Integer insert(E entity);
 
     @InsertSql
-    Long insertBatch(Collection<E> list);
+    Integer insertBatch(Collection<E> list);
 
     @DeleteSql
-    Long delete(K id);
+    Integer delete(K id);
 
     @InsertSql
     Long insertBatchMixture(String tableName, List<E> listParam);
